coingecko client moved to clients
