+++ marketdata/clients/coingecko/models.go
